chore(modules): Organize barrels for tree-shaking
