+++ src/auth.ts
@@ -1,0 +1,2 @@
+export * from './auth/auth';
+export * from './auth/auth.module';
