+++ src/database/firebase_object_observable.spec.ts
@@ -1,7 +1,9 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { FirebaseApp, FirebaseAppConfig, AngularFireModule } from '../angularfire2';
 import { COMMON_CONFIG } from '../test-config';
-import { FirebaseObjectObservable, AngularFireDatabaseModule, AngularFireDatabase } from './index';
+import { AngularFireDatabase } from './database';
+import { AngularFireDatabaseModule } from './database.module';
+import { FirebaseObjectObservable } from './firebase_object_observable';
 import { Observer } from 'rxjs/Observer';
 import { map } from 'rxjs/operator/map';
 import * as firebase from 'firebase/app';
