--- src/database/index.ts
@@ -1,7 +1,0 @@
-export * from './database';
-export * from './firebase_list_factory';
-export * from './firebase_list_observable';
-export * from './firebase_object_factory';
-export * from './firebase_object_observable';
-export * from './query_observable';
-export * from './database.module';
