+++ src/database/query_observable.spec.ts
@@ -2,7 +2,7 @@ import { Observable } from 'rxjs/Observable';
 import { Observer } from 'rxjs/Observer';
 import { Subject } from 'rxjs/Subject';
 import { Query, ScalarQuery } from '../interfaces';
-import { getOrderObservables, observeQuery } from './index';
+import { getOrderObservables, observeQuery } from './query_observable';
 
 function scalarQueryTest(query: ScalarQuery, done: any) {
   const queryObservable = observeQuery(query);
