+++ tsconfig.publish.es6.json
@@ -20,8 +20,8 @@
   },
   "files": [
     "src/index.ts",
-    "src/auth/index.ts",
-    "src/database/index.ts",
+    "src/auth.ts",
+    "src/database.ts",
     "node_modules/zone.js/dist/zone.js.d.ts"
   ],
   "angularCompilerOptions": {
